RESEFINET/requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
cachetools==5.5.1
certifi==2024.8.30
charset-normalizer==3.4.1
cycler==0.11.0
filelock==3.12.2
fonttools==4.53.1
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
geomloss==0.2.6
google-auth==2.38.0
google-auth-oauthlib==0.4.6
grpcio==1.62.3
huggingface-hub==0.16.4
hydra-core==1.3.2
idna==3.10
imageio==2.31.2
imbalanced-learn==0.7.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
kiwisolver==1.4.5
markdown==3.4.4
markupsafe==2.1.5
matplotlib==3.7.5
mkl-service==2.4.0
mkl_fft==1.3.6
mkl_random==1.2.2
multidict==6.0.5
munkres==1.1.4
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
packaging==24.1
pandas==2.0.3
pillow==10.4.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic-core==2.14.6
pyparsing==3.1.4
pytz==2025.1
pywavelets==1.4.1
pyyaml==6.0.1
qudida==0.0.4
regex==2024.4.16
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.7.10
timm==0.9.12
tokenizers==0.13.3
torchmetrics==0.11.4
tornado==6.3.3
tqdm==4.66.5
transformers==4.30.2
typing-extensions==4.7.1
urllib3==1.26.18
wcwidth==0.2.13
werkzeug==2.2.3
yarl==1.9.4
zipp==3.15.0
# 项目核心包
streamlit==1.32.2
python-dotenv==1.0.1
inference-sdk==0.9.0